(REF): Annotators refactoring
