+++ src/main/resources/META-INF/plugin.xml
@@ -92,8 +92,7 @@
         <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustHighlightingAnnotator"/>
         <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustLiteralAnnotator"/>
         <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustExpressionAnnotator"/>
-        <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustItemsAnnotator"/>
-        <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustInvalidSyntaxAnnotator"/>
+        <annotator language="Rust" implementationClass="org.rust.ide.annotator.RustErrorAnnotator"/>
 
         <!-- Line Marker Providers -->
