+++ src/test/resources/org/rust/ide/annotator/fixtures/errors/create_file_and_expand_module_quick_fix/after/foo/mod.rs
