+++ src/test/resources/org/rust/ide/annotator/fixtures/errors/invalid_module_declarations.rs
