+++ src/test/resources/org/rust/ide/annotator/fixtures/errors/invalid_trait_impl_fix.rs
